// src/shared/interfaces/pacientes.interface.ts
export type Pacientes = {
  pac_anticonceptivos_orales?: string | null;
  pac_celular?: string | null;
  pac_contacto_alternativo?: string | null;
  pac_correo?: string | null;
  pac_diabetes?: string | null;
  pac_direccion?: string | null;
  pac_dispositivo_intrauterino?: string | null;
  pac_eps_id?: 1;
  pac_estado_civil?: string | null;
  pac_estrato?: string | null;
  pac_fecha_nacimiento?: string | null;
  pac_fuma?: string | null;
  pac_infecciones_ts?: string | null;
  pac_menopausia?: string | null;
  pac_nivel_educacion?: string | null;
  pac_parejas_sexuales?: string | null;
  pac_partos?: string | null;
  pac_per_identificacion?: string | null;
  pac_peso?: string | null;
  pac_primera_mestruacion?: string | null;
  pac_prueba_ADN_VPH?: string | null;
  pac_regimen_salud?: string | null;
  pac_relacion_condon?: string | null;
  pac_situacion_laboral?: string | null;
  pac_talla?: string | null;
  pac_telefono?: string | null;
  pac_telefono_contacto_alternativo?: string | null;
  pac_tiempo_insercion_DIU?: string | null;
  pac_ultima_citologia?: string | null;
  pac_vacuna_vph?: string | null;
};